Move ZipkinThriftConverter to jaeger-zipkin
